requirements.txt
opencv-python
dlib
face_recognition
pyserial
pip install -r requirements.txt
